Change Type of ResourceTypeProcessor
Change Type of ResourceTypeProcessor from int to int64 - correcting build issues generated from https://github.com/vmware/go-vcloud-director/pull/355

+++ types/v56/constants.go
@@ -137,7 +137,7 @@ const (
 
 const (
 	ResourceTypeOther     int = 0
-	ResourceTypeProcessor int = 3
+	ResourceTypeProcessor int64 = 3
 	ResourceTypeMemory    int = 4
 	ResourceTypeIDE       int = 5
 	ResourceTypeSCSI      int = 6
